Added Integration With STMD
I ended up having to copy the files directly into the the extensions directory
because ruby's `mkmf` library has no way of building files in nested
directories. There must be a better way to do this, but this made it work.

This is also not working for UTF8 characters.  My guess is that there is some
problem when translating ruby strings to the strings that the library uses.